auto init list

+++ darkmoon.cpp
@@ -521,7 +521,7 @@ void checkotherwm(void)
 
 void cleanup(void)
 {
-	Arg a = {.ui = ~0};
+	Arg a = {.ui = (unsigned int)~0};
 	Layout foo = {"", NULL};
 	Monitor *m;
 	size_t i;
